Move `_FormInfoMap` and `_FormInfo` to `shuup.core.utils.forms`
These classes have re-usability value, thus they are now removed to
a more sane place.

Make Service Provider views to use these new classes.
